@@ -50,32 +61,39 @@
     "prepare": "husky install"
   },
   "devDependencies": {
-    "@babel/core": "7.21.0",
-    "@babel/helper-module-imports": "7.18.6",
+    "@babel/core": "7.21.4",
+    "@babel/helper-module-imports": "7.21.4",
     "@babel/plugin-proposal-class-properties": "7.18.6",
     "@babel/plugin-proposal-object-rest-spread": "7.20.7",
     "@babel/plugin-proposal-optional-chaining": "7.21.0",
     "@babel/plugin-transform-object-assign": "7.18.6",
-    "@babel/preset-env": "7.20.2",
-    "@babel/preset-typescript": "7.21.0",
+    "@babel/preset-env": "7.21.4",
+    "@babel/preset-typescript": "7.21.4",
     "@babel/register": "7.21.0",
-    "@microsoft/api-documenter": "7.21.5",
+    "@microsoft/api-documenter": "7.21.7",
     "@microsoft/api-extractor": "7.34.4",
+    "@rollup/plugin-alias": "5.0.0",
+    "@rollup/plugin-babel": "6.0.3",
+    "@rollup/plugin-commonjs": "24.1.0",
+    "@rollup/plugin-node-resolve": "15.0.2",
+    "@rollup/plugin-replace": "5.0.2",
+    "@rollup/plugin-terser": "0.4.1",
+    "@rollup/plugin-typescript": "11.1.0",
     "@types/chai": "4.3.4",
     "@types/chart.js": "2.9.37",
     "@types/mocha": "10.0.1",
     "@types/sinon-chai": "3.2.9",
-    "@typescript-eslint/eslint-plugin": "5.54.1",
-    "@typescript-eslint/parser": "5.54.1",
+    "@typescript-eslint/eslint-plugin": "5.59.2",
+    "@typescript-eslint/parser": "5.59.2",
     "babel-loader": "9.1.2",
     "babel-plugin-transform-remove-console": "6.9.4",
     "chai": "4.3.7",
     "chart.js": "2.9.4",
-    "chromedriver": "110.0.0",
-    "coverage-istanbul-loader": "3.0.5",
+    "chromedriver": "112.0.0",
     "doctoc": "2.2.1",
-    "eslint": "8.35.0",
-    "eslint-config-prettier": "8.7.0",
+    "es-check": "7.1.1",
+    "eslint": "8.39.0",
+    "eslint-config-prettier": "8.8.0",
     "eslint-plugin-import": "2.27.5",
     "eslint-plugin-mocha": "10.1.0",
     "eslint-plugin-node": "11.1.0",
@@ -84,32 +102,31 @@
     "http-server": "14.1.1",
     "husky": "8.0.3",
     "jsonpack": "1.1.5",
-    "karma": "6.4.1",
-    "karma-chrome-launcher": "3.1.1",
-    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma": "6.4.2",
+    "karma-chrome-launcher": "3.2.0",
+    "karma-coverage": "2.2.0",
     "karma-mocha": "2.0.1",
     "karma-mocha-reporter": "2.2.5",
+    "karma-rollup-preprocessor": "7.0.8",
     "karma-sinon-chai": "2.0.2",
     "karma-sourcemap-loader": "0.4.0",
-    "karma-webpack": "5.0.0",
-    "lint-staged": "13.1.4",
+    "lint-staged": "13.2.1",
     "markdown-styles": "3.2.0",
     "micromatch": "4.0.5",
     "mocha": "10.2.0",
-    "netlify-cli": "12.14.0",
-    "node-fetch": "3.3.0",
-    "prettier": "2.8.4",
+    "node-fetch": "3.3.1",
+    "npm-run-all": "4.1.5",
+    "prettier": "2.8.8",
     "promise-polyfill": "8.3.0",
+    "rollup": "3.21.0",
+    "rollup-plugin-istanbul": "4.0.0",
     "sauce-connect-launcher": "1.3.2",
-    "selenium-webdriver": "4.8.1",
-    "semver": "7.3.8",
-    "sinon": "15.0.1",
+    "selenium-webdriver": "4.9.0",
+    "semver": "7.5.0",
+    "sinon": "15.0.4",
     "sinon-chai": "3.7.0",
-    "typescript": "4.9.5",
+    "typescript": "5.0.4",
     "url-toolkit": "2.2.5",
-    "webpack": "5.75.0",
-    "webpack-cli": "5.0.1",
-    "webpack-dev-server": "4.11.1",
-    "webpack-merge": "5.8.0"
+    "wrangler": "2.17.0"
   }
 }
