@@ -11,8 +11,7 @@ import {
   appendFrame,
 } from '../../../src/demux/adts';
 import { ErrorTypes } from '../../../src/errors';
-
-const sinon = require('sinon');
+import sinon from 'sinon';
 
 describe('getAudioConfig', function () {
   it('should trigger a MEDIA_ERROR event if sample index is invalid', function () {
