@@ -41,17 +41,12 @@ try {
     // remove v
     intermediateVersion = intermediateVersion.substring(1);
 
-    const suffix =
-      process.env.NETLIFY && process.env.CONTEXT === 'deploy-preview'
-        ? `pr.${
-            process.env.REVIEW_ID /* set by netlify */
-          }.${getCommitHash().slice(0, 8)}`
-        : process.env.NETLIFY && process.env.CONTEXT === 'branch-deploy'
-        ? `branch.${process.env.BRANCH /* set by netlify */.replace(
-            /[^a-zA-Z0-9]/g,
-            '-'
-          )}.${getCommitHash().slice(0, 8)}`
-        : `0.canary.${getCommitNum()}`;
+    const suffix = process.env.CF_PAGES
+      ? `pr.${process.env.CF_PAGES_BRANCH.replace(
+          /[^a-zA-Z-]/g,
+          '-'
+        )}.${getCommitHash().slice(0, 8)}`
+      : `0.canary.${getCommitNum()}`;
 
     newVersion = `${intermediateVersion}${isStable ? '-' : '.'}${suffix}`;
   }
