@@ -1,3 +1,3 @@
 // This file is inserted as a shim for modules which we do not want to include into the distro.
-// This replacement is done in the "resolve" section of the webpack config.
+// This replacement is done in the "alias" plugin of the rollup config.
 module.exports = undefined;
