@@ -71,7 +71,9 @@ export function enableLogs(debugConfig: boolean | ILogger, id: string): void {
     // Some browsers don't allow to use bind on console object anyway
     // fallback to default if needed
     try {
-      exportedLogger.log(`Debug logs enabled for "${id}"`);
+      exportedLogger.log(
+        `Debug logs enabled for "${id}" in hls.js version ${__VERSION__}`
+      );
     } catch (e) {
       exportedLogger = fakeLogger;
     }
