@@ -8,13 +8,15 @@ import type {
   LoaderResponse,
 } from '../types/loader';
 import { LoadStats } from '../loader/load-stats';
-import { RetryConfig } from '../config';
+import { type HlsConfig, RetryConfig } from '../config';
 import { getRetryDelay, shouldRetry } from './error-helper';
 
 const AGE_HEADER_LINE_REGEX = /^age:\s*[\d.]+\s*$/im;
 
 class XhrLoader implements Loader<LoaderContext> {
-  private xhrSetup: Function | null;
+  private xhrSetup:
+    | ((xhr: XMLHttpRequest, url: string) => Promise<void> | void)
+    | null;
   private requestTimeout?: number;
   private retryTimeout?: number;
   private retryDelay: number;
@@ -25,20 +27,20 @@ class XhrLoader implements Loader<LoaderContext> {
   private loader: XMLHttpRequest | null = null;
   public stats: LoaderStats;
 
-  constructor(config /* HlsConfig */) {
-    this.xhrSetup = config ? config.xhrSetup : null;
+  constructor(config: HlsConfig) {
+    this.xhrSetup = config ? config.xhrSetup || null : null;
     this.stats = new LoadStats();
     this.retryDelay = 0;
   }
 
-  destroy(): void {
+  destroy() {
     this.callbacks = null;
     this.abortInternal();
     this.loader = null;
     this.config = null;
   }
 
-  abortInternal(): void {
+  abortInternal() {
     const loader = this.loader;
     self.clearTimeout(this.requestTimeout);
     self.clearTimeout(this.retryTimeout);
@@ -52,7 +54,7 @@ class XhrLoader implements Loader<LoaderContext> {
     }
   }
 
-  abort(): void {
+  abort() {
     this.abortInternal();
     if (this.callbacks?.onAbort) {
       this.callbacks.onAbort(this.stats, this.context, this.loader);
@@ -63,7 +65,7 @@ class XhrLoader implements Loader<LoaderContext> {
     context: LoaderContext,
     config: LoaderConfiguration,
     callbacks: LoaderCallbacks<LoaderContext>
-  ): void {
+  ) {
     if (this.stats.loading.start) {
       throw new Error('Loader can only be used once.');
     }
@@ -74,7 +76,7 @@ class XhrLoader implements Loader<LoaderContext> {
     this.loadInternal();
   }
 
-  loadInternal(): void {
+  loadInternal() {
     const { config, context } = this;
     if (!config) {
       return;
@@ -89,16 +91,18 @@ class XhrLoader implements Loader<LoaderContext> {
     if (xhrSetup) {
       Promise.resolve()
         .then(() => {
+          if (this.stats.aborted) return;
           return xhrSetup(xhr, context.url);
         })
-        .catch((error) => {
+        .catch((error: Error) => {
           xhr.open('GET', context.url, true);
           return xhrSetup(xhr, context.url);
         })
         .then((data) => {
+          if (this.stats.aborted) return;
           this.openAndSendXhr(xhr, context, config, data);
         })
-        .catch((error) => {
+        .catch((error: Error) => {
           // IE11 throws an exception on xhr.open if attempting to access an HTTP resource over HTTPS
           this.callbacks!.onError(
             { code: xhr.status, text: error.message },
@@ -142,6 +146,7 @@ class XhrLoader implements Loader<LoaderContext> {
       }
 
       const headers = this.context.headers;
+      const { maxTimeToFirstByteMs, maxLoadTimeMs } = config.loadPolicy;
       if (headers) {
         for (const header in headers) {
           xhr.setRequestHeader(header, headers[header]);
@@ -160,15 +165,18 @@ class XhrLoader implements Loader<LoaderContext> {
       xhr.responseType = context.responseType as XMLHttpRequestResponseType;
       // setup timeout before we perform request
       self.clearTimeout(this.requestTimeout);
+      config.timeout =
+        maxTimeToFirstByteMs && Number.isFinite(maxTimeToFirstByteMs)
+          ? maxTimeToFirstByteMs
+          : maxLoadTimeMs;
       this.requestTimeout = self.setTimeout(
         this.loadtimeout.bind(this),
-        config.loadPolicy.maxTimeToFirstByteMs
+        config.timeout
       );
       xhr.send();
     }
   }
-
-  readystatechange(): void {
+  readystatechange() {
     const { context, loader: xhr, stats } = this;
     if (!context || !xhr) {
       return;
@@ -189,13 +197,15 @@ class XhrLoader implements Loader<LoaderContext> {
           stats.loading.start
         );
         // readyState >= 2 AND readyState !==4 (readyState = HEADERS_RECEIVED || LOADING) rearm timeout as xhr not finished yet
-        self.clearTimeout(this.requestTimeout);
-        config.timeout = config.loadPolicy.maxLoadTimeMs;
-        this.requestTimeout = self.setTimeout(
-          this.loadtimeout.bind(this),
-          config.loadPolicy.maxLoadTimeMs -
-            (stats.loading.first - stats.loading.start)
-        );
+        if (config.timeout !== config.loadPolicy.maxLoadTimeMs) {
+          self.clearTimeout(this.requestTimeout);
+          config.timeout = config.loadPolicy.maxLoadTimeMs;
+          this.requestTimeout = self.setTimeout(
+            this.loadtimeout.bind(this),
+            config.loadPolicy.maxLoadTimeMs -
+              (stats.loading.first - stats.loading.start)
+          );
+        }
       }
 
       if (readyState === 4) {
@@ -218,6 +228,8 @@ class XhrLoader implements Loader<LoaderContext> {
           const len =
             xhr.responseType === 'arraybuffer' ? data.byteLength : data.length;
           stats.loaded = stats.total = len;
+          stats.bwEstimate =
+            (stats.total * 8000) / (stats.loading.end - stats.loading.first);
           if (!this.callbacks) {
             return;
           }
@@ -255,7 +267,7 @@ class XhrLoader implements Loader<LoaderContext> {
     }
   }
 
-  loadtimeout(): void {
+  loadtimeout() {
     const retryConfig = this.config?.loadPolicy.timeoutRetry;
     const retryCount = this.stats.retry;
     if (shouldRetry(retryConfig, retryCount, true)) {
@@ -292,7 +304,7 @@ class XhrLoader implements Loader<LoaderContext> {
     );
   }
 
-  loadprogress(event: ProgressEvent): void {
+  loadprogress(event: ProgressEvent) {
     const stats = this.stats;
 
     stats.loaded = event.loaded;
