@@ -1,3 +1,45 @@
 import 'promise-polyfill/src/polyfill';
-const testsContext = require.context('./unit', true);
-testsContext.keys().forEach(testsContext);
+import './unit/hls';
+import './unit/events';
+import './unit/controller/abr-controller';
+import './unit/controller/audio-stream-controller';
+import './unit/controller/audio-track-controller';
+import './unit/controller/base-stream-controller';
+import './unit/controller/buffer-controller-operations';
+import './unit/controller/buffer-controller';
+import './unit/controller/buffer-operation-queue';
+import './unit/controller/cap-level-controller';
+import './unit/controller/cmcd-controller';
+import './unit/controller/content-steering-controller';
+import './unit/controller/eme-controller';
+import './unit/controller/error-controller';
+import './unit/controller/ewma-bandwidth-estimator';
+import './unit/controller/fragment-finders';
+import './unit/controller/fragment-tracker';
+import './unit/controller/gap-controller';
+import './unit/controller/latency-controller';
+import './unit/controller/level-controller';
+import './unit/controller/level-helper';
+import './unit/controller/stream-controller';
+import './unit/controller/subtitle-stream-controller';
+import './unit/controller/subtitle-track-controller';
+import './unit/controller/timeline-controller-nonnative';
+import './unit/controller/timeline-controller';
+import './unit/crypt/aes-decryptor';
+import './unit/demuxer/adts';
+import './unit/demuxer/base-audio-demuxer';
+import './unit/demuxer/id3';
+import './unit/demuxer/transmuxer';
+import './unit/loader/date-range';
+import './unit/loader/fragment-loader';
+import './unit/loader/fragment';
+import './unit/loader/level';
+import './unit/loader/playlist-loader';
+import './unit/utils/attr-list';
+import './unit/utils/binary-search';
+import './unit/utils/buffer-helper';
+import './unit/utils/discontinuities';
+import './unit/utils/exp-golomb';
+import './unit/utils/output-filter';
+import './unit/utils/texttrack-utils';
+import './unit/utils/vttparser';
