@@ -1,9 +1,9 @@
 import BufferOperationQueue from '../../../src/controller/buffer-operation-queue';
-import { BufferOperation, SourceBuffers } from '../../../src/types/buffer';
+import type { BufferOperation, SourceBuffers } from '../../../src/types/buffer';
 
-import * as sinon from 'sinon';
-import * as chai from 'chai';
-import * as sinonChai from 'sinon-chai';
+import sinon from 'sinon';
+import chai from 'chai';
+import sinonChai from 'sinon-chai';
 
 chai.use(sinonChai);
 const expect = chai.expect;
