@@ -2,7 +2,7 @@ import CMCDController from '../../../src/controller/cmcd-controller';
 import HlsMock from '../../mocks/hls.mock';
 import type { CMCDControllerConfig } from '../../../src/config';
 
-import * as chai from 'chai';
+import chai from 'chai';
 
 const expect = chai.expect;
 
