@@ -1,5 +1,5 @@
-import * as chai from 'chai';
-import * as sinonChai from 'sinon-chai';
+import chai from 'chai';
+import sinonChai from 'sinon-chai';
 import EwmaBandWidthEstimator from '../../../src/utils/ewma-bandwidth-estimator';
 
 chai.use(sinonChai);
