@@ -14,8 +14,8 @@ import type {
   FragLoadedData,
 } from '../../../src/types/events';
 
-import * as chai from 'chai';
-import * as sinonChai from 'sinon-chai';
+import chai from 'chai';
+import sinonChai from 'sinon-chai';
 
 chai.use(sinonChai);
 const expect = chai.expect;
@@ -667,5 +667,6 @@ function createMockFragment(
       data.endPTS
     );
   });
+  frag.relurl = 'not-frag-hint';
   return frag;
 }
