@@ -8,14 +8,10 @@ import { Events } from '../../../src/events';
 describe('SubtitleTrackController', function () {
   let subtitleTrackController;
   let videoElement;
-  let hls;
   let sandbox;
-  let fake;
 
   beforeEach(function () {
-    sandbox = sinon.createSandbox();
-    fake = sinon.useFakeXMLHttpRequest();
-    hls = new Hls({
+    const hls = new Hls({
       renderNatively: true,
     });
 
@@ -53,18 +49,25 @@ describe('SubtitleTrackController', function () {
 
     const textTrack1 = videoElement.addTextTrack('subtitles', 'English', 'en');
     const textTrack2 = videoElement.addTextTrack('subtitles', 'Swedish', 'se');
+    const textTrack3 = videoElement.addTextTrack(
+      'captions',
+      'Untitled CC',
+      'en'
+    );
+
     textTrack1.groupId = 'default-text-group';
     textTrack2.groupId = 'default-text-group';
+    textTrack3.groupId = 'default-text-group';
     subtitleTrackController.groupId = 'default-text-group';
 
     textTrack1.mode = 'disabled';
     textTrack2.mode = 'disabled';
+    textTrack3.mode = 'disabled';
+    sandbox = sinon.createSandbox();
   });
 
   afterEach(function () {
-    hls.destroy();
     sandbox.restore();
-    fake.restore();
   });
 
   describe('onTextTrackChanged', function () {
@@ -93,6 +96,16 @@ describe('SubtitleTrackController', function () {
       subtitleTrackController.onTextTracksChanged();
 
       expect(subtitleTrackController.subtitleTrack).to.equal(0);
+    });
+
+    it('should set subtitleTrack id captions track is showing', function () {
+      expect(subtitleTrackController.subtitleTrack).to.equal(-1);
+
+      videoElement.textTracks[2].mode = 'showing';
+      subtitleTrackController.onTextTracksChanged();
+
+      expect(videoElement.textTracks[2].kind).to.equal('captions');
+      expect(subtitleTrackController.subtitleTrack).to.equal(2);
     });
   });
 
