@@ -22,8 +22,8 @@ describe('TimelineController', function () {
         Events.SUBTITLE_TRACKS_UPDATED,
         {
           subtitleTracks: [
-            { id: 0, name: 'en' },
-            { id: 1, name: 'ru' },
+            { id: 0, name: 'en', attrs: { LANGUAGE: 'en', NAME: 'en' } },
+            { id: 1, name: 'ru', attrs: { LANGUAGE: 'ru', NAME: 'ru' } },
           ],
         }
       );
@@ -41,8 +41,8 @@ describe('TimelineController', function () {
         Events.SUBTITLE_TRACKS_UPDATED,
         {
           subtitleTracks: [
-            { id: 0, name: 'en' },
-            { id: 1, name: 'ru' },
+            { id: 0, name: 'en', attrs: { LANGUAGE: 'en', NAME: 'en' } },
+            { id: 1, name: 'ru', attrs: { LANGUAGE: 'ru', NAME: 'ru' } },
           ],
         }
       );
@@ -62,8 +62,8 @@ describe('TimelineController', function () {
 
       timelineController.onSubtitleTracksUpdated(Events.MANIFEST_LOADED, {
         subtitleTracks: [
-          { id: 0, name: 'en' },
-          { id: 1, name: 'ru' },
+          { id: 0, name: 'en', attrs: { LANGUAGE: 'en', NAME: 'en' } },
+          { id: 1, name: 'ru', attrs: { LANGUAGE: 'ru', NAME: 'ru' } },
         ],
       });
 
@@ -78,8 +78,8 @@ describe('TimelineController', function () {
 
       timelineController.onSubtitleTracksUpdated(Events.MANIFEST_LOADED, {
         subtitleTracks: [
-          { id: 0, name: 'ru' },
-          { id: 1, name: 'en' },
+          { id: 0, name: 'ru', attrs: { LANGUAGE: 'ru', NAME: 'ru' } },
+          { id: 1, name: 'en', attrs: { LANGUAGE: 'en', NAME: 'en' } },
         ],
       });
 
@@ -146,6 +146,7 @@ describe('TimelineController', function () {
           {
             id: 0,
             name: 'en',
+            attrs: {},
           },
         ],
       });
