@@ -1,7 +1,7 @@
 import { DateRange } from '../../../src/loader/date-range';
 import { AttrList } from '../../../src/utils/attr-list';
 
-import * as chai from 'chai';
+import chai from 'chai';
 
 const expect = chai.expect;
 
